Match load_directory loop types to num_entries (uint32_t)
MSVC /W4 flags 'i <= pak->num_entries' (int vs uint32_t) as a
signed/unsigned mismatch; under /WX that's a hard build error.
entry_pos comes from the same uint32_t arithmetic, so promote both.
fseek still takes long — cast explicitly.

+++ src/pakfile.c
@@ -125,13 +125,13 @@ void load_pakfile(Pak_t *pak) {
 void load_directory(Pak_t *pak) {
     Pakfileentry_t *current = NULL;
     Pakfileentry_t *last = NULL;
-    int i, entry_pos;
+    uint32_t i, entry_pos;
 
     if (pak->num_entries > 0) {
         for (i = 1; i <= pak->num_entries; i++) {
-            entry_pos = pak->diroffset + pak->dirlength 
+            entry_pos = pak->diroffset + pak->dirlength
                 - (i * PAKFILE_DIR_ENTRY_SIZE);
-            fseek(fp, entry_pos, SEEK_SET);
+            fseek(fp, (long)entry_pos, SEEK_SET);
 
             current = calloc(1, sizeof(Pakfileentry_t));
             if (fread(current->filename, PAKFILE_PATH_MAX, 1, fp) != 1
